+++ src/system/esp_ll_stm32f723e_discovery.c
@@ -311,8 +311,8 @@ configure_uart(uint32_t baudrate) {
  * \param[in]       len: Number of bytes to send
  * \return          Number of bytes sent
  */
-static uint16_t
-send_data(const void* data, uint16_t len) {
+static size_t
+send_data(const void* data, size_t len) {
     uint16_t i;
     const uint8_t* d = data;
